Replace magic numbers (-1, -2) in multiplayer lobby code with named constants (inline functions). Get rid of several static_casts. Fix "src/network/nethost.cc:717: warning: comparison between signed and unsigned integer expressions". Fix const correctness and style issues.
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@4680 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/network/netclient.h
@@ -79,7 +79,7 @@ struct NetClient :
 	virtual void setPlayerInit    (uint8_t number, uint8_t index);
 	virtual void setPlayerName    (uint8_t number, std::string const & name);
 	virtual void setPlayer        (uint8_t number, PlayerSettings ps);
-	virtual void setPlayerNumber  (int32_t number);
+	virtual void setPlayerNumber  (uint8_t number);
 	virtual void setPlayerReady   (uint8_t number, bool ready);
 	virtual bool getPlayerReady   (uint8_t number);
